+++ nequip/model/__init__.py
@@ -1,5 +1,10 @@
 from ._eng import EnergyModel, SimpleIrrepsConfig
-from ._grads import ForceOutput, PartialForceOutput, StressForceOutput
+from ._grads import (
+    ForceOutput,
+    PartialForceOutput,
+    StressForceOutput,
+    ParaStressForceOutput,
+)
 from ._scaling import RescaleEnergyEtc, PerSpeciesRescale
 from ._weight_init import (
     uniform_initialize_FCs,
@@ -18,6 +23,7 @@ __all__ = [
     ForceOutput,
     PartialForceOutput,
     StressForceOutput,
+    ParaStressForceOutput,
     RescaleEnergyEtc,
     PerSpeciesRescale,
     uniform_initialize_FCs,